implemented function that will return all trains available from db and function that will check the username if it has the correct password

+++ src/main/java/Trains/Train.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Train {
-    private long trainId;
+    private int trainId;
     private String departureLocation;
     private String arrivalLocation;
     private int seatsAvailable;
     private int seatsTotal;
 
-    public Train(long trainId, String departureLocation, String arrivalLocation, int seatsAvailable, int seatsTotal) {
+    public Train(int trainId, String departureLocation, String arrivalLocation, int seatsAvailable, int seatsTotal) {
         this.trainId = trainId;
         this.departureLocation = departureLocation;
         this.arrivalLocation = arrivalLocation;
